Update GeoTiff values
Increased the values that encode information in the GeoTIFF. This is to be able to inspect the files visually with a simple tiff viewer. For the protected areas for example, when protected areas are encoded with 1, but the value range is up to 255, then you won't be able to see the visually inspect the areas.

This inverts the previous commit, but that is fine at this moment.