Fix allocation of availability to regions
So far I only counted the number of pixels with a certain availability to derive a fraction of land that is available. From there I used the overall region size to derive the available land in km^2. This approach is wrong, because the used projection WGS84 is not equi-distant. Pixels in the north are smaller than pixels in the south. Instead, now I am using EPSG:3035 to derive the size of each pixel and determine the overall available land from there.

Also, I am determining the available land now in one step insted of two as before.