Add allocation of exlusive economic zones to administrative regions
Using a simple approach in which shares of offshore eligibility are allocated to all regions that share a coast with the EEZ. The share is proportional to the length of the shared coast.

Also using a simple approach to determine the length of the shared coast; it's fairly rough but accurate enough for this analysis.