Refactor to break linear workflow pt 1/2
Currently there is a long sequential workflow in which one attribute after the other is added to the regions shapes. That is not necessary, as instead several things can be calculated in parallel, and whenever necessary, all needed attributes can simply be merged.

This is only partly done. Currently the workflow cannot be run in its entirety.

This work is preparation for adding protected areas as another scenario dimension. Without this work, loads of calculation would be run more than once unnecessarily.