Fix population in water bodies
Some people had been allocated to water bodies. I am now identifying water bodies with two criteria: a flag in the lau2 data that indicates that a particular region ist not a true commune (whatever that means, metadata missing). And the water to area ratio. I then remove people from these water bodies without reallocating them. I do not reallocate because the fraction of population for which this happens is very small: < 0.1%.