Fix border gap between Montenegro and Serbia
The fix is a simple visiaul patch, not a proper solution. I believe the issue is a political one: the demarcation process hasn't been finished back in 2013 when the NUTS/LAU data has been produced. Maybe by now the process is finished or has proceeded and that is visible in the new NUTS/LAU 2016 data. That data hasn't been released yet though.