Remove pv farms from farm land
There is a conflict in land use when using pv on farm land. It is possible to grow crops on the same land with pv installations but this not standard yet and the implications e.g. on efficiency on both uses aren't clear. I will consider using farm land for pv, but not in the default case. It really should be a scenario.