Move to JRC population data set
The dataset is close to the settlement layer that I am using as it is based on the same inputs. It is also based on the GPW dataset I was using before and in that way is a refinement. Also, its resolution is 250m and thus higher than ~1km from before.