Add effect of roof orientation to roof mounted PV
To properly account for roof orientation, I will need a statistical model of roof orientation (have that already) plus simulations at least on NUTS2 level for all roof types. I don't have that at the moment. To get closer to real numbers, I used the following approach:

From sonnendach data I know that 20% of all roofs are each N,S,E,W-facing or flat. A quick round of simulations in Germany told me that N-facing has 50% of S-facing and E/W-facing has 80% of S-facing. Assuming PV on flat roofs is southfacing as well, this gives a total capacity factor reduction of 2/5*1 + 2/5*0.8 + 1/5*0.5 = 0.82.

As the simplest way of appplying the factor, I added it to the installable power per squaremeter where it obviously does not belong. Will change everyting as soon as I have the simulations.