Update sensitivity analysis and analyse results
* improved performance of sensitivity analysis
* model has several outputs now: urban pop, rural pop, and number regions
* add two more scenarios to show the importance of rooftop pv

The result is in any case: rooftop pv is overwhelmingly important. Why is that? The technical potential of rooftop pv is a good bit lower than for pv farms, onshore and offshore wind seperately.

I guess it's this:
In most regions, rooftop pv alone is able to fullfil the demand. This is different to other technologies, which in total have a higher potential, but their spatial distribution is different. What makes rooftop pv so successfull is the fact that is so close to the consumer. So, a value close to 100% availability allows to reduce the number of unsatisfied population to nearly zero.