Add coarse estimation of actual rooftop pv potential
So far I used the sonnendach data which estimates the theoretic potential of roof mounted pv. In "Berechnung des Potentials für Gemeinden" a method is described to derive the actual usuable potential based on the theoretic potential. Roughly it means that 70% of the theoretic potential is available. I applied this factor here.

The correct method relies on the type of each building, which is defined in the Swiss GWR data set. If I am able to get this data set, I will implement the method correctly.