Add DEGURBA degree of urbanisation
This is based on UN/OECD/EC approach to classify the urbanisation of regions. I manually checked it and it seems to have many issues. There are several "urban" areas with low population density and population and with visually rural structure (e.g. Germigny-l'Évêque or Chambry
77910) and "rural" areas with high population density and population (e.g. Gentilly next to Paris).

Still it might be useful to compare my population density only approach to this official classification.