Fix the estimation of rooftop area
Now that the full sonnendach.ch data set is available, it is easy to compare to their results. Here, I made corrections to the way sonnendach.ch is used to determine the correction factor applied to ESM data.

These changes cause a 30% reduction of installable capacity for roof mounted PV and hence, a 30% decrease in the potential of roof mounted PV. This makes local autarky more often impossible than before.