Increase max slope for open field pv
The source says the range is from 3-10°. Because we can only show the impossibility of autarky, it makes sense to go for an optimistic value. It is also consistent with the 20° for wind, which are rather optimistic as well.

This change increases the continental energy yield quite a bit, and also shifts from onshore wind to open field pv. It doesn't change a lot in the final results though, mainly because it increases the energy yield, where normed potential was already high before the change.