Fix necessary land estimation
Fix two things:

(1) Build on protected areas. This is important, otherwise many regions have no open field potential. This is a regression.
(2) Handle those cases, where no open field potential exist. This typically is the case in urban municipalities. In the previous draft, I set fraction of land necessary to 1 for all of them. However, that's not exactly true: an urban municipality may have large amounts of unbuilt land that are on hand hand not available for renewable generation (because they are parks for example), but maybe only a fraction of them is necessary. Instead, I am not using a national metric for how much energy can be generated on the unbuilt land, and then determine the fraction of unbuilt land necessary. The results are qualitatively the same.